new analyze.ts testing with mini workspaces
